Move the parsing logic to parser.hpp
