+++ src/CMakeLists.txt
@@ -9,6 +9,7 @@ add_executable(anzu
                stack_frame.cpp
                op_codes.cpp
                lexer.cpp
+               parser.cpp
                object.cpp
                functions.cpp)
